Move dynamic content folder
