+++ src/extend/dynamic-content/attributes.js
