+++ src/extend/dynamic-content/hoc/withContentLink.js
@@ -1,4 +1,3 @@
-
 import { select } from '@wordpress/data';
 
 function EmptyLink( { text } ) {
