+++ src/extend/dynamic-content/inspector-controls/CommentsControl.js
