+++ src/extend/dynamic-content/inspector-controls/DynamicSourceControl.js
