+++ src/extend/dynamic-content/inspector-controls/PostMetaControl.js
@@ -1,8 +1,8 @@
 import usePostRecord from '../hooks/usePostRecord';
 import { __ } from '@wordpress/i18n';
-import AdvancedSelect from '../../../../../components/advanced-select';
+import AdvancedSelect from '../../../components/advanced-select';
 
-export default function AuthorMetaControl( props ) {
+export default function PostMetaControl( props ) {
 	const {
 		isActive = false,
 		postType,
@@ -14,17 +14,17 @@ export default function AuthorMetaControl( props ) {
 	const record = usePostRecord( postType, postId );
 	let options = [];
 
-	if ( record && record.author && record.author.meta ) {
+	if ( record && record.meta ) {
 		options = Object
-			.keys( record.author.meta )
+			.keys( record.meta )
 			.map( ( metaKey ) => ( { value: metaKey, label: metaKey } ) );
 	}
 
 	// ACF support
-	if ( record && record.author && record.author.acf ) {
+	if ( record && record.acf ) {
 		options = options.concat(
 			Object
-				.keys( record.author.acf )
+				.keys( record.acf )
 				.map( ( metaKey ) => ( { value: metaKey, label: metaKey } ) )
 		);
 	}
@@ -33,13 +33,13 @@ export default function AuthorMetaControl( props ) {
 		<>
 			{ isActive &&
 				<AdvancedSelect
-					id={ 'gblocks-select-author-meta-control' }
-					label={ __( 'Author meta field', 'generateblocks' ) }
-					placeholder={ __( 'Author meta field', 'generateblocks' ) }
+					id={ 'gblocks-select-post-meta-control' }
+					label={ __( 'Post meta field', 'generateblocks' ) }
+					placeholder={ __( 'Post meta field', 'generateblocks' ) }
 					options={ options }
 					value={ { value: metaFieldName, label: metaFieldName } }
 					isSearchable
-					isLoading={ ( !record && ! record.author && ! record.author.meta ) }
+					isLoading={ ( ! record && ! record.meta ) }
 					onChange={ ( option ) => {
 						setAttributes( { metaFieldName: option.value } );
 					} }
