+++ src/extend/dynamic-content/utils/applyContext.js
@@ -1,9 +1,9 @@
 /**
  * Applies the context to the attributes.
  *
- * @param {Object} context The block context.
+ * @param {Object} context    The block context.
  * @param {Object} attributes The block attributes.
- * @returns {Object} The attributes.
+ * @return {Object} The attributes.
  */
 export default function applyContext( context, attributes ) {
 	if ( attributes.dynamicSource === 'current-post' ) {
